ai-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
spacy>=3.7.0
textblob>=0.17.1
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
setuptools>=65.0.0
beautifulsoup4>=4.12.0
googlesearch-python>=1.2.3

